# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: FixingApp Backend API
  description: |
    Backend MVP API for connecting employers with workers.
    
    This API provides endpoints for:
    - Authentication and user management
    - Job posting and management
    - Job applications
    - Worker verification
    - Reviews and ratings
    - Complaints handling
    - Admin functions
    
    **Authentication**: Most endpoints require a JWT Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fixingapp.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3000/api
    description: API base path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Not Found"
        message: "Route GET /api/invalid not found"
...
paths:
...
  /api/workers/certificates/status:
    get:
      tags:
        - Workers
      summary: Get certificate status
      description: Get certificate verification status (Worker only)
      operationId: getCertificateStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Certificate status
          content:
            application/json:
              schema:
                type: object
                properties:
                  isVerified:
                    type: boolean
                  certificates:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        imageUrl:
                          type: string
                        status:
                          type: string
                          enum: [PENDING, APPROVED, REJECTED]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/complaints:
    post:
      tags:
        - Complaints
      summary: File a complaint
      description: Submit a complaint about a job
      operationId: fileComplaint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jobId
                - reason
              properties:
                jobId:
                  type: integer
                  example: 1
                reason:
                  type: string
                  example: "Worker did not show up on time"
                evidences:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [IMAGE, LOG]
                      url:
                        type: string
            example:
              jobId: 1
              reason: "Worker did not show up on time"
              evidences:
                - type: "IMAGE"
                  url: "/static/evidence1.jpg"
      responses:
        '201':
          description: Complaint filed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  jobId:
                    type: integer
                  reason:
                    type: string
                  status:
                    type: string
                    enum: [PENDING]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/complaints/my:
    get:
      tags:
        - Complaints
      summary: Get my complaints
      description: Get all complaints filed by the authenticated user
      operationId: getMyComplaints
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    jobId:
                      type: integer
                    reason:
                      type: string
                    status:
                      type: string
                      enum: [PENDING, RESOLVED]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/certificates/pending:
    get:
      tags:
        - Admin
      summary: Get pending certificates
      description: Get all certificates pending verification (Admin only)
      operationId: getPendingCertificates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of pending certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    workerId:
                      type: integer
                    imageUrl:
                      type: string
                    status:
                      type: string
                      enum: [PENDING]
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/complaints/{complaintId}/resolve:
    post:
      tags:
        - Admin
      summary: Resolve a complaint
      description: Resolve a complaint with a decision (Admin only)
      operationId: resolveComplaint
      security:
        - bearerAuth: []
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
              properties:
                decision:
                  type: string
                  enum: [ACCEPT, REJECT]
                  example: "ACCEPT"
                notes:
                  type: string
                  example: "Complaint is valid"
            example:
              decision: "ACCEPT"
              notes: "Complaint is valid"
      responses:
        '200':
          description: Complaint resolved successfully
        '400':
          description: Bad request (invalid decision)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'